requirements.txt
configparser==3.5.0
diff-match-patch==20121119
Django==1.10.2
django-appconf==1.0.2
django-autofixture==0.12.1
django-ckeditor==5.1.1
django-dynamic-preferences==0.8.2
django-extensions==1.7.4
django-fortune==1.1.1
django-import-export==0.5.1
django-robots==2.0
djangorestframework==3.4.7
enum34==1.1.6
flake8==3.0.4
fortune==1.0
grizzled-python==1.0.7
gunicorn==19.6.0
Markdown==2.6.7
mccabe==0.5.2
pathlib2==2.1.0
pep8==1.7.0
persisting-theory==0.2.1
Pillow==3.4.1
pycodestyle==2.0.0
pyflakes==1.2.3
rcssmin==1.0.6
rjsmin==1.0.12
setupfiles==0.0.13
six==1.10.0
tablib==0.11.2
